Adding AuthModule and SharedModule, refactor login components
Added a new module called `AuthModule` which is loaded by `AppModule`. It contains 4 components, 3 of which are partially complete:
- login
- register
- forgot-password
- reset-password

A new module called `SharedModule` was added and is used by _both_ `AppModule` and `AuthModule`.  It contains all of the stuff we'll most likely need on many components to simplify dependency injection.

Some new npm packages were added for `ngDynamicForms`. We are using the `ui-ng-bootstrap` UI package to work with the `NgbModule` that was already installed. These modules are configured and imported/exported in the `SharedModule`. Components currently using it are: 
- `auth/login`
- `auth/forgot-password`
- `auth/register`
... soon to include `auth/reset-password` as well.